Load 3.3.559 into branch/prcs-import-33.
git-svn-id: svn://svn.bearnip.com/ldmud/branch/prcs-import-33@1189 56160b6e-f8f8-0310-bd54-bccebfe6a2f2

+++ src/efuns.c
@@ -6492,7 +6492,7 @@ v_member (svalue_t *sp, int num_arg)
             efun_arg_error(2, T_NUMBER, sp->type, sp);
         str = sp[-1].u.str;
         
-        if (hasStart && startpos >= mstrsize(str))
+        if (hasStart && (size_t)startpos >= mstrsize(str))
             i = -1;
         else
         {
@@ -6713,7 +6713,7 @@ v_rmember (svalue_t *sp, int num_arg)
         if (sp->type != T_NUMBER)
             efun_arg_error(2, T_NUMBER, sp->type, sp);
         str = sp[-1].u.str;
-        if (!hasStart || startpos >= mstrsize(str))
+        if (!hasStart || (size_t)startpos >= mstrsize(str))
             startpos = mstrsize(str);
         i = sp->u.number;
         if ((i & ~0xff) != 0)
